added notifications page  and started adding stream

+++ app/api/posts/[postId]/likes/route.ts
@@ -1,6 +1,7 @@
 import { currentUser } from "@/lib/actions/session.actions";
 import prisma from "@/lib/prisma";
-import { LikeInfo } from "@/lib/types";
+import { getPostDataInclude, LikeInfo } from "@/lib/types";
+import { NotificationType } from "@prisma/client";
 
 export async function GET(req: Request, { params: { postId } }: { params: { postId: string } }) {
 
@@ -53,20 +54,52 @@ export async function POST(req: Request, { params: { postId } }: { params: { pos
         const loggedInUser = await currentUser();
         if (!loggedInUser) return Response.json({ error: "Unauthorized" }, { status: 401 });
 
-        await prisma.like.upsert({
+        const post = await prisma.post.findUnique({
             where: {
-                userId_postId: {
-                    postId,
-                    userId: loggedInUser.id as string
-                }
+                id: postId,
             },
-            create: {
-                userId: loggedInUser.id as string,
-                postId,
-            },
-            update: {}
+            select: {
+                userId: true
+            }
+        });
 
-        })
+        if (!post) return Response.json({ error: "Post not found" }, { status: 404 });
+
+
+        await prisma.$transaction([
+            prisma.like.upsert({
+                where: {
+                    userId_postId: {
+                        postId,
+                        userId: loggedInUser.id as string
+                    }
+                },
+                create: {
+                    userId: loggedInUser.id as string,
+                    postId,
+                },
+                update: {}
+
+            }),
+
+            ...(loggedInUser.id !== post.userId ?
+
+                [
+                    prisma.notification.create({
+                        data: {
+                            type: NotificationType.LIKE,
+                            issuerId: loggedInUser.id as string,
+                            recipientId: post.userId as string,
+                            postId
+                        }
+                    })
+                ] :[])
+        ])
+
+
+
+
+
         return new Response()
     } catch (error) {
         console.log("[POST_LIKES_ROUTE]", error);
@@ -80,12 +113,35 @@ export async function DELETE(req: Request, { params: { postId } }: { params: { p
     try {
         const loggedInUser = await currentUser();
         if (!loggedInUser) return Response.json({ error: "Unauthorized" }, { status: 401 });
-        await prisma.like.deleteMany({
+        const post = await prisma.post.findUnique({
             where: {
-                userId: loggedInUser.id as string,
-                postId
+                id: postId,
+            },
+            select: {
+                userId: true
             }
-        })
+        });
+
+        if (!post) return Response.json({ error: "Post not found" }, { status: 404 });
+
+        await prisma.$transaction([
+            prisma.like.deleteMany({
+                where: {
+                    userId: loggedInUser.id as string,
+                    postId
+                }
+            }),
+
+            prisma.notification.deleteMany({
+                where : {
+                    postId,
+                    issuerId : loggedInUser.id as string,
+                    recipientId  : post.userId as string  ,
+                    type : NotificationType.LIKE
+                }
+            })
+        ])
+      
         return new Response()
     } catch (error) {
         console.log("[DELETE_LIKES_ROUTE]", error);
